pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conscious"
version = "1.0.0"
description = "Conscious - Your Local AI Voice Companion"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "moshi>=0.2.12",
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    "huggingface-hub>=0.20.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "colorama>=0.4.6",
]

[project.optional-dependencies]
memory = [
    "mem0ai>=0.1.0",
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.2.0",
]
emotion = [
    "transformers>=4.36.0",
    "librosa>=0.10.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]
all = [
    "conscious[memory,emotion,dev]",
]

[project.scripts]
conscious = "conscious.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"conscious.data" = ["*.yaml"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
